add @root/request, the 80/20 request alternative
I wrote this several months ago based on the request.js docs and behavior. It's a drop-in replacement with ZERO dependencies. I feel like it's worth including because it's only 500 LoC and it, like I said, it's a simple drop-in replacement for 80%+ of what people actually use request.js for.